% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_partitioning.R
\name{add_user_tagging}
\alias{add_user_tagging}
\title{Adds pixel tagging to mark pixels selected by the user as object & background}
\usage{
add_user_tagging(input_image_df, limits_object, limits_background)
}
\arguments{
\item{input_image_df}{df storing image in a specific format}

\item{limits_object}{list with coordinates of area (rectangle) selected as part of the object}

\item{limits_background}{list with coordinates of area (rectangle) selected as part of the background}
}
\value{
df with additional column named user_tagging
}
\description{
Adds pixel tagging to mark pixels selected by the user as object & background
}
\examples{
NULL
}
