% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_partitioning.R
\name{calc_params_distribution}
\alias{calc_params_distribution}
\title{Calculates parameters of the estimated distribution}
\usage{
calc_params_distribution(input_image_df, type, method, n_dist = NA)
}
\arguments{
\item{input_image_df}{df storing image in a specific format}

\item{type}{string, "object" or "background"}

\item{method}{selected method of estimating object and background color distributions}

\item{n_dist}{number of distributions selected}
}
\value{
named list with parameters of estimated distribution
}
\description{
Calculates parameters of the estimated distribution
}
\examples{
NULL
}
