% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_partitioning.R
\name{calc_vertices_probs}
\alias{calc_vertices_probs}
\title{Calculates probabilities that a pixel color is coming from the same
distribution as pixels selected as object / as background}
\usage{
calc_vertices_probs(input_image_df, method, object_params, background_params)
}
\arguments{
\item{input_image_df}{df storing image in a specific format}

\item{method}{selected method of estimating object and background color distributions}

\item{object_params}{list with parameters of the estimated object distribution}

\item{background_params}{list with parameters of the estimated background distribution}
}
\value{
df with additional columns named dnorm_object & dnorm_background
}
\description{
Calculates probabilities that a pixel color is coming from the same
distribution as pixels selected as object / as background
}
\examples{
NULL
}
