% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{conv_image_to_df}
\alias{conv_image_to_df}
\title{Converts cimg image to a df with a specific format}
\usage{
conv_image_to_df(image)
}
\arguments{
\item{image}{object of class cimg}
}
\value{
df, each row representing one pixel
}
\description{
Converts cimg image to a df with a specific format
}
\examples{
NULL
}
