% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_partitioning.R
\name{conv_image_to_graph}
\alias{conv_image_to_graph}
\title{Converts data frame with edges and capacities to igraph format}
\usage{
conv_image_to_graph(edges_caps)
}
\arguments{
\item{edges_caps}{df with edges and capacities}
}
\value{
igraph object
}
\description{
Converts data frame with edges and capacities to igraph format
}
\examples{
NULL
}
