% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_partitioning.R
\name{create_partitioning}
\alias{create_partitioning}
\title{Creates image partitioning into object and background based on selected image areas}
\usage{
create_partitioning(
  input_image_df,
  limits_object,
  limits_background,
  method,
  n_dist = NULL
)
}
\arguments{
\item{input_image_df}{df storing image in a specific format}

\item{limits_object}{list with coordinates of area (rectangle) selected as part of the object}

\item{limits_background}{list with coordinates of area (rectangle) selected as part of the background}

\item{method}{selected method of estimating object and background color distributions}

\item{n_dist}{list with number of distributions selected for method "mixed" for object & background}
}
\value{
partitioning, two lists of pixels, containing pixels classified as object and as background
}
\description{
Creates image partitioning into object and background based on selected image areas
}
\examples{
NULL
}
