% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_image.R
\name{cut_image}
\alias{cut_image}
\title{Performs an image segmentation}
\usage{
cut_image(image, method, k = NULL)
}
\arguments{
\item{image}{a 3-dimensional input image (the range of the pixel values should be preferably in the range 0 to 255)}

\item{method}{a character string specifying the variation of graph cut method. It can be either "regular", "mixed or "mixed_bic"}

\item{k}{}
}
\value{

}
\description{
Performs an image segmentation
}
