% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{get_image}
\alias{get_image}
\title{Loads the image from specified path and resized it uniformly}
\usage{
get_image(image_path, max_size = 80)
}
\arguments{
\item{image_path}{path to the image}

\item{max_size}{specifies maximum number of rows and columns}
}
\value{
object of cimg class, resized to not exceed max_size on any of the dimensions
}
\description{
Loads the image from specified path and resized it uniformly
}
\examples{
NULL
}
