% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_partitioning.R
\name{get_params_list}
\alias{get_params_list}
\title{Creates a list of parameters for further calculations}
\usage{
get_params_list(input_image_df, method, n_dist)
}
\arguments{
\item{input_image_df}{df storing image in a specific format}

\item{method}{selected method of estimating object and background color distributions}

\item{n_dist}{list with number of distributions selected for method "mixed" for object & background}
}
\value{
named list with parameters
}
\description{
Creates a list of parameters for further calculations
}
\examples{
NULL
}
