% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_segmentation_app}
\alias{run_segmentation_app}
\title{The main menu for the mixedgraphcut package}
\usage{
run_segmentation_app()
}
\description{
This function opens a Shiny app with a menu that will allow the user to run the different image segmentations.
}
\details{
Run this function with no arguments to start the main menu (a Shiny app) for interactive image segmentation.
}
\examples{
\dontrun{run_segmentation_app()}
}
\author{
Marta Gajewska
}
